requirements.txt
fastapi
uvicorn
tensorflow-cpu
opencv-python-headless
pillow
numpy
python-multipart